requirements.txt
Django==5.2.7
djangorestframework==3.15.2
python-dotenv==1.0.1
requests==2.32.3
gunicorn==21.2.0
psycopg
whitenoise==6.8.2
twilio==9.2.3
pypdf==4.2.0
lxml==5.2.2
zeep==4.2.1
cryptography==42.0.8
reportlab==4.2.2
weasyprint==61.2
